TypedGraphBuilder implements GraphBuilder
