+++ src/dachshund/typed_graph_builder.rs
@@ -7,15 +7,53 @@
 extern crate fxhash;
 
 use crate::dachshund::error::{CLQError, CLQResult};
-use crate::dachshund::graph_builder::GraphBuilder;
-use crate::dachshund::id_types::NodeId;
-use crate::dachshund::node::{Node, NodeEdge};
+use crate::dachshund::graph_base::GraphBase;
+use crate::dachshund::graph_builder_base::GraphBuilderBase;
+use crate::dachshund::id_types::{GraphId, NodeTypeId, NodeId};
+use crate::dachshund::node::{Node, NodeBase, NodeEdge};
 use crate::dachshund::row::EdgeRow;
 use crate::dachshund::typed_graph::TypedGraph;
 use fxhash::FxHashMap;
+use std::collections::{HashMap, HashSet};
 
-pub struct TypedGraphBuilder {}
-impl GraphBuilder<TypedGraph> for TypedGraphBuilder {
+pub struct TypedGraphBuilder {
+    pub min_degree: Option<usize>,
+    pub graph_id: GraphId,
+}
+impl GraphBuilderBase for TypedGraphBuilder {
+
+    type GraphType = TypedGraph;
+    type RowType = EdgeRow;
+
+    fn from_vector(&self, rows: &Vec<EdgeRow>) -> CLQResult<TypedGraph> {
+        let mut source_ids: HashSet<NodeId> = HashSet::new();
+        let mut target_ids: HashSet<NodeId> = HashSet::new();
+        let mut target_type_ids: HashMap<NodeId, NodeTypeId> = HashMap::new();
+        for r in rows.iter() {
+            assert!(self.graph_id == r.graph_id);
+            source_ids.insert(r.source_id);
+            target_ids.insert(r.target_id);
+            target_type_ids.insert(r.target_id, r.target_type_id);
+        }
+
+        // warrant a canonical order on the id vectors
+        let mut source_ids_vec: Vec<NodeId> = source_ids.into_iter().collect();
+        source_ids_vec.sort();
+        let mut target_ids_vec: Vec<NodeId> = target_ids.into_iter().collect();
+        target_ids_vec.sort();
+
+        let mut node_map: FxHashMap<NodeId, Node> =
+            Self::init_nodes(&source_ids_vec, &target_ids_vec, &target_type_ids);
+        Self::populate_edges(rows, &mut node_map)?;
+        let mut graph = Self::create_graph(node_map, source_ids_vec, target_ids_vec)?;
+        if let Some(min_degree) = self.min_degree {
+            graph = Self::prune(graph, rows, min_degree)?;
+        }
+        Ok(graph)
+    }
+}
+
+impl TypedGraphBuilder {
     fn create_graph(
         nodes: FxHashMap<NodeId, Node>,
         core_ids: Vec<NodeId>,
@@ -78,5 +116,117 @@ impl GraphBuilder<TypedGraph> for TypedGraphBuilder {
             }
         }
         Ok(())
+    }
+
+    // initializes nodes in the graph with empty neighbors fields.
+    fn init_nodes(
+        core_ids: &[NodeId],
+        non_core_ids: &[NodeId],
+        non_core_type_ids: &HashMap<NodeId, NodeTypeId>,
+    ) -> FxHashMap<NodeId, Node> {
+        let mut node_map: FxHashMap<NodeId, Node> = FxHashMap::default();
+        for &id in core_ids {
+            let node = Node::new(
+                id,             // node_id,
+                true,           // is_core,
+                None,           // non_core_type,
+                Vec::new(),     // edges,
+                HashMap::new(), //neighbors
+            );
+            node_map.insert(id, node);
+        }
+        for &id in non_core_ids {
+            let node = Node::new(
+                id,                           // node_id,
+                false,                        // is_core,
+                Some(non_core_type_ids[&id]), // non_core_type,
+                Vec::new(),                   // edges,
+                HashMap::new(),               // neighbors
+            );
+            node_map.insert(id, node);
+        }
+        node_map
+    }
+
+
+    /// Trims edges greedily, until all edges in the graph have degree at least min_degree.
+    /// Note that this function does not delete any nodes -- just finds nodes to delete. It is
+    /// called by `prune`, which actually does the deletion.
+    pub fn trim_edges(node_map: &mut FxHashMap<NodeId, Node>, min_degree: &usize) -> HashSet<NodeId> {
+        let mut degree_map: HashMap<NodeId, usize> = HashMap::new();
+        for (node_id, node) in node_map.iter() {
+            let node_degree: usize = node.degree();
+            degree_map.insert(*node_id, node_degree);
+        }
+        let mut nodes_to_delete: HashSet<NodeId> = HashSet::new();
+        loop {
+            let mut nodes_to_update: HashSet<NodeId> = HashSet::new();
+            for (node_id, node_degree) in degree_map.iter() {
+                if node_degree < min_degree && !nodes_to_delete.contains(node_id) {
+                    nodes_to_update.insert(*node_id);
+                    nodes_to_delete.insert(*node_id);
+                }
+            }
+            if nodes_to_update.is_empty() {
+                break;
+            }
+            for node_id in nodes_to_update.iter() {
+                let node: &Node = &node_map[node_id];
+                for n in node.edges.iter() {
+                    let neighbor_node_id: NodeId = n.target_id;
+                    let current_degree: usize = degree_map[&neighbor_node_id];
+                    degree_map.insert(neighbor_node_id, current_degree - 1);
+                }
+            }
+        }
+        nodes_to_delete
+    }
+    
+    /// Takes an already-built graph and the edge rows used to create it, returning a
+    /// new graph, where all nodes are assured to have degree at least min_degree.
+    /// The provision of a <Self as GraphBuilderBase>::GraphType is necessary, since the notion of "degree" does
+    /// not make sense outside of a graph.
+    pub fn prune(graph: <Self as GraphBuilderBase>::GraphType, rows: &[EdgeRow], min_degree: usize) -> CLQResult<<Self as GraphBuilderBase>::GraphType> {
+        let mut target_type_ids: HashMap<NodeId, NodeTypeId> = HashMap::new();
+        for r in rows.iter() {
+            target_type_ids.insert(r.target_id, r.target_type_id);
+        }
+        let (filtered_source_ids, filtered_target_ids, filtered_rows) =
+            Self::get_filtered_sources_targets_rows(graph, min_degree, rows);
+        let mut filtered_node_map: FxHashMap<NodeId, Node> =
+            Self::init_nodes(&filtered_source_ids, &filtered_target_ids, &target_type_ids);
+        Self::populate_edges(&filtered_rows, &mut filtered_node_map)?;
+        Self::create_graph(filtered_node_map, filtered_source_ids, filtered_target_ids)
+    }
+    /// called by `prune`, finds source and target nodes to exclude, as well as edges to exclude
+    /// when rebuilding the graph from a filtered vector of `EdgeRows`.
+    fn get_filtered_sources_targets_rows(
+        mut graph: <Self as GraphBuilderBase>::GraphType,
+        min_degree: usize,
+        rows: &[EdgeRow],
+    ) -> (Vec<NodeId>, Vec<NodeId>, Vec<EdgeRow>) {
+        let exclude_nodes: HashSet<NodeId> = Self::trim_edges(graph.get_mut_nodes(), &min_degree);
+        let filtered_source_ids: Vec<NodeId> = graph
+            .get_core_ids()
+            .iter()
+            .filter(|x| !exclude_nodes.contains(x))
+            .cloned()
+            .collect();
+        let filtered_target_ids: Vec<NodeId> = graph
+            .get_non_core_ids()
+            .unwrap()
+            .iter()
+            .filter(|x| !exclude_nodes.contains(x))
+            .cloned()
+            .collect();
+        let filtered_rows: Vec<EdgeRow> = rows
+            .iter()
+            .filter(|x| {
+                !(exclude_nodes.contains(&x.source_id) || (exclude_nodes.contains(&x.target_id)))
+            })
+            .cloned()
+            .collect();
+        // todo: make member of struct
+        (filtered_source_ids, filtered_target_ids, filtered_rows)
     }
 }
