+++ tests/pruning.rs
@@ -7,7 +7,6 @@
 extern crate lib_dachshund;
 use lib_dachshund::dachshund::candidate::Candidate;
 use lib_dachshund::dachshund::error::{CLQError, CLQResult};
-use lib_dachshund::dachshund::graph_builder::GraphBuilder;
 use lib_dachshund::dachshund::id_types::{GraphId, NodeId};
 use lib_dachshund::dachshund::test_utils::{
     assert_nodes_have_ids, gen_test_transformer, process_raw_vector,
@@ -39,7 +38,7 @@ fn simple_test(raw: Vec<String>, min_degree: usize, expected_len: usize) -> CLQR
     let rows = process_raw_vector(&transformer, raw)?;
 
     let mut graph: TypedGraph =
-        transformer.build_pruned_graph::<TypedGraphBuilder, TypedGraph>(graph_id, &rows)?;
+        transformer.build_pruned_graph(graph_id, &rows)?;
     let exclude_nodes: HashSet<NodeId> =
         TypedGraphBuilder::trim_edges(&mut graph.nodes, &min_degree);
     assert_eq!(exclude_nodes.len(), expected_len);
@@ -99,13 +98,13 @@ fn test_prune_small_clique() -> CLQResult<()> {
     let transformer = gen_test_transformer(ts, "author".to_string())?;
     let rows = process_raw_vector(&transformer, raw)?;
     let mut graph: TypedGraph =
-        transformer.build_pruned_graph::<TypedGraphBuilder, TypedGraph>(graph_id, &rows)?;
+        transformer.build_pruned_graph(graph_id, &rows)?;
     assert_eq!(graph.nodes.len(), 5);
     graph = TypedGraphBuilder::prune(graph, &rows, 2)?;
     assert_eq!(graph.nodes.len(), 4);
     let v = Vec::new();
     let res: Candidate<TypedGraph> = transformer
-        .process_graph::<TypedGraph>(&graph, &v, graph_id, true)?
+        .process_graph(&graph, &v, graph_id, true)?
         .top_candidate;
     assert_nodes_have_ids(&graph, &res.core_ids, vec![1, 2], true);
     assert_nodes_have_ids(&graph, &res.non_core_ids, vec![3, 4], false);
@@ -154,10 +153,10 @@ fn test_full_prune_small_clique() -> CLQResult<()> {
     let rows_prune = process_raw_vector(&transformer_prune, raw.clone())?;
 
     let graph: TypedGraph = transformer_prune
-        .build_pruned_graph::<TypedGraphBuilder, TypedGraph>(graph_id, &rows_prune)?;
+        .build_pruned_graph(graph_id, &rows_prune)?;
     let v_prune = Vec::new();
     let result_prune = transformer_prune
-        .process_clique_rows::<TypedGraphBuilder, TypedGraph>(
+        .process_clique_rows(
             &graph,
             &v_prune,
             graph_id,
@@ -189,10 +188,10 @@ fn test_full_prune_small_clique() -> CLQResult<()> {
     let rows = process_raw_vector(&transformer, raw)?;
 
     let graph: TypedGraph =
-        transformer.build_pruned_graph::<TypedGraphBuilder, TypedGraph>(graph_id, &rows)?;
+        transformer.build_pruned_graph(graph_id, &rows)?;
     let v = Vec::new();
     let result = transformer
-        .process_clique_rows::<TypedGraphBuilder, TypedGraph>(&graph, &v, graph_id, false, &sender)?
+        .process_clique_rows(&graph, &v, graph_id, false, &sender)?
         .ok_or_else(CLQError::err_none)?;
     sender.send((None, true)).unwrap();
     let candidate = result.top_candidate;
